Add hack to fix flickering issue
Occasionally there was a small flicker when placing buildings. This
seemed to be caused by the preview icon becoming momentarily
visible. This in turn seemed to be caused by a bug in Godot that a
mouse_exit signal is emitted when the mouse moves to a child
control (in this case the building just being placed).

We now check whether the mouse has actually left the area before
marking it as outside.