Refactor that also fixes tooltip disappearance bug
This is a small refactor of the code base helping also to fix the
'tooltip disappearance' bug. The building panel used to completely
refresh when a building's info_updated signal was emitted, and if at
that moment the user was trying to read a stat label tooltip, it would
disappear. In late game, these updates are very frequent and make
reading the tooltip almost impossible.

Some Godot warnings were also fixed and some string literals are now
replaced with enums.