Fix building popping out behavior
Since we added some free space around building sprite (for drawing
outlines), the popping out behavior is not acting very fluently and
there were also issues with placing a building right next to the other
one. We now add a transparent "main area" over the sprite which is
used for handling mouse events.