Fix the "shrink on click" bug
We had a bug that when we moved over a building (causing a to pop out)
and then we clicked on the building, the building would receive a
mouse_exit event and shrink back, even though the mouse cursor was
still over it.

This was caused by a rather complicated bug in Godot. What happens is
that when we click on the building, we add some widgets to the
building panel. Some of these widgets have their rect_size set so that
the controls would theoretically expand over the building, even though
the container would not allow that, but apparently Godot recognized
this as another control covering the building, causing a mouse_exit
event to be sent.

This only happens when the added widgets are after the building in the
tree, so another option might have been to restructure the nodes so
that the widgets were added somewhere before the building in the
tree. However fixing margins and sizes in the BuidlingStat and
BuildingWidget scenes seemed like a more robust solution.

Notice that the sizes and margins were automatically set by Godot
somehow before, so I'm doing this, hoping Godot doesn't decide to
change them again later, causing this bug to reappear.