Make planet rotation with the wheel smoother
I had to do some tricks to make this work. In particular, checking
`Input.is_action_pressed` does not seem to work, since the button
release event seems to immediately happen after pressed, which means
in _physics_processed the is_action_pressed function always returns
false.

So now we set acceleration in the _unhandled_input function and use a
timer to reset it after a while.