Disable collision shape instead of using monitorable
For disabling preview icon "collisions" we now disable/enable the
area's CollisionShape2D (using the "disabled" property), instead of
setting monitorable to false/true.

This has the advantage of not needing the hack to make it work
immediately on existing collisions as soon as it's changed, so we
don't need the hack we had for that purpose.

Using `set_deferred` to set this property is in accordance with the
docs.