Fix smoke reset issue
It used to be that whenever we changed the rate of the Smoke node, it
would reset emitting, causing all current smoke to vanish. This is now
fixed by using multiple particle emitters. When we change the rate, we
replace the main particle emitter with a new one, leaving the old one
in place until it runs out.