Change BuildingPanel type to Panel
When BuildingPanel's type was PanelContainer, we would have this bug
that the settings button and bottom panel would suddenly disappear.

The reason this happened was that we would add some widgets to the
building panel and remove others, but the removal is not instant and
so the parent PanelContainer's size would grow, but it won't shrink
back. I think that is because we have a filler control in the toolbox
that also grows and so the top-level hud/hbox would become permanently
larger, moving the settings button and the bottom container
off-screen.

Changing BuildingPanel's type to Panel (instead of PanelContainre)
would mean that it won't change size even if its children grow too
big.