Fix wrong action price being used
Since the latest changes, we consume money after the action is
performed. But this was causing the game to use the price after the
action being performed, which could be different than the original
price. We now calculate the price first, perform the action and then
consume the original price.